Elaborate on `useCount` usage for Extra Credit 1
This could just be my own misunderstanding, but when reading the original text:

> It should throw an error indicating that `useCount` must be used within a CountProvider.

I took this to mean that if you *must* use `useCount` within a `CountProvider`, and an error should be thrown if a call to `useCount` is not found, instead of `useCount` *can only be used* from within a `CountProvider`. 

I realise now this wouldn't have made much sense, but hopefully this can save someone else from needing to look at the answer to understanding what they need(ed) to do!